Move example and fix tests when database and tables don't exist
