+++ example/app.js
@@ -1,58 +1,68 @@
+'use strict';
+
 const rethink = require('rethinkdbdash');
 const feathers = require('feathers');
 const rest = require('feathers-rest');
 const socketio = require('feathers-socketio');
 const bodyParser = require('body-parser');
 const service = require('../lib');
-
-// Connect to a local RethinkDB server.
 const r = rethink({
   db: 'feathers'
 });
 
+r.db('feathers').tableList().contains('todos')
+  .do(function(tableExists) {
+    return r.branch(
+      tableExists, {
+        created: 0
+      },
+      r.db('feathers').tableCreate('todos')
+    );
+  }).run();
+
+let counter = 0;
+const todoService = service({
+  Model: r,
+  name: 'todos',
+  paginate: {
+    default: 2,
+    max: 4
+  }
+}).extend({
+  _find(params) {
+    params = params || {};
+    params.query = params.query || {};
+    if (!params.query.$sort) {
+      params.query.$sort = {
+        counter: 1
+      };
+    }
+
+    return this._super(params);
+  },
+
+  create(data, params) {
+    data.counter = ++counter;
+    return this._super(data, params);
+  }
+});
+
 // Create a feathers instance.
-const app = feathers()
-  // Enable the REST provider for services.
+let app = feathers()
+  // Enable REST services
   .configure(rest())
-  // Enable the socketio provider for services.
+  // Enable Socket.io services
   .configure(socketio())
   // Turn on JSON parser for REST services
   .use(bodyParser.json())
   // Turn on URL-encoded parser for REST services
   .use(bodyParser.urlencoded({
     extended: true
-  }));
-
-// Create your database if it doesn't exist.
-r.dbList().contains('feathers')
-  .do(dbExists => r.branch(dbExists, {
-    created: 0
-  }, r.dbCreate('feathers'))).run()
-  // Create the table if it doesn't exist.
-  .then(() =>
-    r.dbList().contains('messages')
-    .do(tableExists => r.branch(
-      tableExists, {
-        created: 0
-      },
-      r.dbCreate('messages'))).run()
-  )
-  // Create and register a Feathers service.
-  .then(() => {
-    app.use('messages', service({
-      Model: r,
-      x
-      name: 'messages',
-      paginate: {
-        default: 10,
-        max: 50
-      }
-    }));
-  })
-  .catch(err => console.log(err));
+  }))
+  .use('/todos', todoService);
 
 // Start the server.
-var port = 3030;
-app.listen(port, function() {
-  console.log(`Feathers server listening on port ${port}`);
-});
+const port = 3030;
+
+// require('http').createServer();
+module.exports = app.listen(port);
